Add image integration tests
We're adding integration tests that exercise the image migration
handler.

The tests will use temporary Openstack credentials, purging the
resources at the end of the run. The credentials are provided
through module scoped fixtures, so tests belonging to the same
module will reuse the same set of credentials.

See the readme for sample invocations.