Allow migrating volumes and volume types (#12)
This PR adds migration handlers for Cinder volumes and volume types.
The user can use a config option to specify whether the volume type
should be preserved or if the default one should be used instead.

In some cases, the original volume type may not be applicable on the
destination side. For this reason, we'll allow the users to manually
recreate the type and register the external migration using
sunbeam-migrate register-external so that subsequent volume migrations
may use it.

There are multiple ways in which the volume data can be migrated,
however we're using the easiest and most widely supported approach:
temporary Glance images. See the inline comments for more details.

The integration tests also use temporary Glance image to validate the
transferred volume contents.

While at it, we're adding a flake8 job. It catches programming errors
that ruff/mypy do not. We'll keep the ruff check though since it has
a slightly different set of rules and we want the code base to match
the coding style of Sunbeam.