Add Manila migration handlers (#16)
This PR adds migration handlers for Manila shares and share types.

The Openstack SDK doensn't support share types yet, so we'll have to use python-manilaclient instead.

Share migrations consist in:

* migrate the share type if preserve_share_type is set
* create the share on the destination cloud, matching the specifications of the original share
* create temporary access rules so that the local host may access the share
* if no local share address is configured, it will be automatically detected based on the local routes
* mount both shares and transfer the data, preserving extended attributes, symlinks, timestamps and ownership information
* remove temporary access rules and mounts

Note that only NFS shares are supported for the time being (same as Sunbeam).

Cursor was used to generate a significant amount of boilerplate code.